third_party/WebKit/LayoutTests/inspector/tracing/frame-model-expected.txt

Test the frames are correctly built based on trace events

Test: main thread only
{
    cpuTime : 4.001
    duration : 16
    endTime : 1016.001
    idle : false
    startTime : 1000.001
    timeByCategory : {
        other : 0.004
        scripting : 3.997
    }
}
{
    cpuTime : 12.994
    duration : 16
    endTime : 1032.001
    idle : false
    startTime : 1016.001
    timeByCategory : {
        other : 3.998
        rendering : 5.999
        scripting : 2.997
    }
}
{
    cpuTime : 1.999
    duration : 16
    endTime : 1048.001
    idle : false
    startTime : 1032.001
    timeByCategory : {
        other : 1.999
    }
}
Test: impl-side only
{
    cpuTime : 0
    duration : 30
    endTime : 1030
    idle : false
    startTime : 1000
    timeByCategory : {
    }
}
{
    cpuTime : 0
    duration : 16
    endTime : 1046
    idle : false
    startTime : 1030
    timeByCategory : {
    }
}
{
    cpuTime : 2.997
    duration : 32
    endTime : 1078
    idle : false
    startTime : 1046
    timeByCategory : {
        painting : 2.997
    }
}
{
    cpuTime : 0
    duration : 3
    endTime : 1081
    idle : false
    startTime : 1078
    timeByCategory : {
    }
}
Test: impl-side with commits
{
    cpuTime : 32.999
    duration : 46
    endTime : 1046
    idle : false
    startTime : 1000
    timeByCategory : {
        other : 3.003
        painting : 1.999
        rendering : 10
        scripting : 17.997
    }
}
{
    cpuTime : 11.999
    duration : 16
    endTime : 1062
    idle : false
    startTime : 1046
    timeByCategory : {
        other : 11.988003
        painting : 0.010997
    }
}
Test: impl-side with interleaving commits
{
    cpuTime : 12.999
    duration : 14.001
    endTime : 1014.001
    idle : false
    startTime : 1000
    timeByCategory : {
        other : 12.986003
        painting : 0.012997
    }
}
{
    cpuTime : 12.999
    duration : 16
    endTime : 1030.001
    idle : false
    startTime : 1014.001
    timeByCategory : {
        other : 0.003
        painting : 3.999
        scripting : 8.997
    }
}
{
    cpuTime : 12.999
    duration : 16
    endTime : 1046.001
    idle : false
    startTime : 1030.001
    timeByCategory : {
        other : 0.003
        painting : 7.999
        scripting : 4.997
    }
}
Test: pre-frame time accounting
{
    cpuTime : 2.999
    duration : 3.001
    endTime : 1035.001
    idle : false
    startTime : 1032
    timeByCategory : {
        other : 1.601
        painting : 0.598
        rendering : 0.8
    }
}
{
    cpuTime : 0
    duration : 28
    endTime : 1063.001
    idle : false
    startTime : 1035.001
    timeByCategory : {
    }
}
{
    cpuTime : 20.998
    duration : 8.001
    endTime : 1071.002
    idle : false
    startTime : 1063.001
    timeByCategory : {
        other : 6.005
        painting : 2.998
        rendering : 2.998
        scripting : 8.997
    }
}
{
    cpuTime : 8.998
    duration : 10
    endTime : 1081.002
    idle : false
    startTime : 1071.002
    timeByCategory : {
        other : 5.002
        painting : 0.998
        rendering : 2.998
    }
}
{
    cpuTime : 0
    duration : 15
    endTime : 1096.002
    idle : false
    startTime : 1081.002
    timeByCategory : {
    }
}
Test: record processing order
{
    cpuTime : 15.999
    duration : 13.005
    endTime : 1013.005
    idle : false
    startTime : 1000
    timeByCategory : {
        other : 0.005
        painting : 3.997
        scripting : 11.997
    }
}
{
    cpuTime : 0
    duration : 16.996
    endTime : 1030.001
    idle : false
    startTime : 1013.005
    timeByCategory : {
    }
}
{
    cpuTime : 31.999
    duration : 32
    endTime : 1062.001
    idle : false
    startTime : 1030.001
    timeByCategory : {
        other : 1.005
        painting : 19.997
        scripting : 10.997
    }
}
{
    cpuTime : 0
    duration : 18
    endTime : 1080.001
    idle : false
    startTime : 1062.001
    timeByCategory : {
    }
}
Test: commits without activation
{
    cpuTime : 0
    duration : 14.001
    endTime : 1014.001
    idle : false
    startTime : 1000
    timeByCategory : {
    }
}
{
    cpuTime : 0
    duration : 16
    endTime : 1030.001
    idle : false
    startTime : 1014.001
    timeByCategory : {
    }
}
{
    cpuTime : 0
    duration : 16
    endTime : 1046.001
    idle : false
    startTime : 1030.001
    timeByCategory : {
    }
}
Test: Idle frames
{
    cpuTime : 12.999
    duration : 14.001
    endTime : 1014.001
    idle : false
    startTime : 1000
    timeByCategory : {
        other : 0.003
        painting : 7.999
        scripting : 4.997
    }
}
{
    cpuTime : 0
    duration : 249.999
    endTime : 1264
    idle : true
    startTime : 1014.001
    timeByCategory : {
    }
}
{
    cpuTime : 24
    duration : 32.001
    endTime : 1296.001
    idle : false
    startTime : 1264
    timeByCategory : {
        other : 0.5
        painting : 0.5
        scripting : 23
    }
}
{
    cpuTime : 0
    duration : 15.999
    endTime : 1312
    idle : false
    startTime : 1296.001
    timeByCategory : {
    }
}
